+++ UserInterfaceGame.cpp
@@ -93,8 +93,8 @@ void UserInterfaceGame::SetFPS(int fps)
 	m_fps = fps;
 }
 
-void UserInterfaceGame::GetMousePosition(int & x,int & y)
+void UserInterfaceGame::GetMousePosition(int16_t & x,int16_t & y)
 {
-	x =(int)(m_mousePosition.i);
-	y =(int)(m_mousePosition.j);
+	x =(m_mousePosition.i);
+	y =(m_mousePosition.j);
 }
