+++ UserInterfaceGameMenu.h
@@ -9,7 +9,7 @@ public:
 
 	void Render(ID3D11DeviceContext* deviceContext, XMFLOAT4X4 viewMatrix, XMFLOAT4X4 projectionMatrix);
 	void Update(XMVECTOR cameraPosItion);
-	void GetMousePosition(int &X, int &Y);
+	void GetMousePosition(int16_t &X, int16_t &Y);
 
 private:
